Have added another class which read the information on the stock buy and sell. Have created an activity that will directs to the browse prediction page.
Things that left to do is  to fix the incorrect time (It's only for the current date / end date and time) . For the stock date, they parsed correctly. The other thing to check is the price increase/decrease for the stocks, there is no API for it. Will discuss this with Don and Emily for this. Also it is important to deal with null pointer exceptions which is not yet dealt with.

Todo: The browse prediction page should contain all the predictions. The coding so far only looked at 2 predictions. Will need to add more to the list as progress goes. Will need to look at the layout design as well.

There are some problems so far such as the ranking information. I couldn't get the ranking info and Don not sure either. Have mailed the site owner and await for the reply.