Working on scroll view of the About us page. Added 2 LinearLayout for the pictures (Not sure if its a good way to do it). Another easy option is to use image instead rather than go through the coding trying to place them next to each other, but the advantage of coding is that it scales according to screen size. If use image instead then it will stretch and won't look nice. Have work on arrange the textView properly so that they are centered and are place in stack like strucutures so that they don't overlap each other. Will continue working on this page where there is a comment section below it.
Scroll Viewing is added in but need to test on the actual phone itself to see if it actually working.

Task left to do: Plan to do the timeout for connection, work on the browse prediction and about us section, check to see if we can actually scroll in the phone. Important to check in different screen size, in portrait and landscape mode as well.