Add a border xml file to use for the message and email box. Worked on the about us page layout and almost done. Need to make the box border around the text fields. There are more things to do here such as the text should be scaled according to the screen size. Got the images today from Emily, and need to use crop those images first before put them here.
Scroll Viewing is added in but need to test on the actual phone itself to see if it actually working.

Task left to do: Plan to do the timeout for connection, work on the browse prediction and about us section, check to see if we can actually scroll in the phone. Important to check in different screen size, in portrait and landscape mode as well.