Worked on cropped the image and replaced with the previous crossed-box. Need to replace the rest of these boxes with the proper images. The next thing to do is to make the text scaled according to screen size and test on the phone. The submit button still do nothing at the moment. This should also be the next task for this page, as well as putting the white background in the middle section where the team is.
To Test: Scroll Viewing is added in but need to test on the actual phone itself to see if it actually working.

Task left to do: Plan to do the timeout for connection, work on the browse prediction and about us section, check to see if we can actually scroll in the phone. Important to check in different screen size, in portrait and landscape mode as well.