Added back the largeheap, the memory seems to be running out when scrolling over the search prediciton page. Noting that the allocated memory do not exceed the max default heap size on the phone but rather there seems to be a spike causing high memory (~52Mb) which is technically smaller than before (~60Mb). There are improvement here and will need time for further investigation.
Looked into the application test java abit. This is something that I need to do as well. Writting test is important so that I know if there is something wrong with the code. Now require a bit of references and learning first because this is the first time I use test in Android.

Last but not least, I have done adding the template for the register page (just a reference to the page with nothing at the moment) and a skeleton structure of the result page which user selects on the stock they want to further display the results.