Fixes the rest of the activity using the onBackPressed method to navigate back to the stack.
Other fixes include changes in the cancel button to clear button to clear selected items. Navigate back from search page will retained the selected item in the browse page in which users can clear with a button. Leaving the cancel button there seems to be really redundant because we have a back soft button on the action bar and we have a back button on the Android phone (iPhone does not have it so it makes sense to have it there but not make sense in Android).

Next commit will be looking into fixing the 'All Contracts' as clicking on any of the predictions not loading anything. Get infor pass to the result page and display them (at the moment all info pass except the JsonArray of buy and sell orders). Need to double check to make sure states changes and reload back properly when changes orientation of all pages. Fix the crash when lost internet connection while retrieve the JSON file.